# Optimized Vector move_towards() - move_towards_ip()